requirements.txt
fastapi
python_dotenv
asyncpg
uvicorn
firebase-admin
pydantic-settings
fastapi-mail
sqlalchemy
pytest
pytest-asyncio
httpx
flake8
pymongo[srv]
